+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 Changelog file for dinero-rs project, a command line application for managing finances.
 ## [0.31.0] - xxx
+## Fixed
+- Read quantities like ```-$0.25```, [bug](https://github.com/frosklis/dinero-rs/issues/126)
 ## [0.30.0] - 2021-08-18
 ## Added
 - Show more info when loading the repl